video-production/requirements.txt
# Core video pipeline dependencies
PyYAML>=6.0.1
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
openai>=1.3.0
requests>=2.31.0

# TTS providers
elevenlabs>=0.2.26
azure-cognitiveservices-speech>=1.31.0

# Audio processing
pyloudnorm>=0.1.1
ffmpeg-python>=0.2.0

# Optional: OpenTimelineIO (requires cmake)
# OpenTimelineIO>=0.17.0

# Optional: Whisper for transcription (choose one)
# openai-whisper>=20231117
# faster-whisper>=0.9.0

# Development and utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
